requirements.txt
http://download.pytorch.org/whl/torch-0.3.0.post4-cp35-cp35m-macosx_10_6_x86_64.whl
torchvision
Django
numpy
matplotlib
